pass missingroutecomponent to router externally
