+++ common/src/helpers/router.tsx
@@ -3,7 +3,6 @@ import UniversalRouter, { Routes } from "universal-router";
 import React, { useState, useEffect, useMemo, ReactElement } from "react";
 import fromPairs from "lodash/fromPairs";
 import toPairs from "lodash/toPairs";
-import NotFound from "../components/NotFound";
 
 const HISTORY = createBrowserHistory();
 
@@ -67,6 +66,7 @@ export interface RouterResult {
 
 export const useRouter = (
   routes: Routes<RouterResult>,
+  missingRouteComponent: ReactElement
 ): React.ReactElement | null => {
   const location = useLocation();
   const router = useMemo(() => new UniversalRouter(routes), [routes]);
@@ -84,8 +84,9 @@ export const useRouter = (
       }
     })
     .catch(async (r) => {
-      setComponent(<NotFound />);
-    });
+      setComponent(missingRouteComponent);
+    })
+   ;
   }, [path, router]);
 
   return component;
